Navigation prototype final
This is the last commit of the navigation prototype. More content needs to be added and or refined to make a complete, fully functional, navigation menu. Changes here will be merged with main branch. Then, create a new branch for landing page build which will be a child component of Home.